Prevent autoconf from automatically puttin -O2 in CXXFLAGS (.C files should be compiled without optimizations).
Move the TooN -I directive in to CPP flags where it should be.

Export a cpu variable to Makefile.in. This takes the form of a known CPU (currently) p3, p4, xeon, pM, opteron, or generic if a specific type is unknown (host_cpu is still defined, but is less specific). This is primarily for the (soon to be included) corner detector where one version of the (portable) C++ code runs fast on a P3 and very slow on a pM, and another version runs fast on a pM.