Added GLWindow.  This is a richer implementation of VideoDisplay that hides the XEvent system from the user.
The GLWindow class makes a window and an associated GL context.  When
requested, events on the queue are either processed with a passed event
handler (with callback virtual methods such as on_key_down and on_resize)
or enumerated in a vector<GLWindow::Event> or GLWindow::EventSummary.

Otherwise the functionality is the same as VideoDisplay.

Feel free to add to the interface; I just got tired of linking against a
separate library in order to have a reasonable display interface.