Added implementation of Young - van Vliet recursive Gaussian filters for float images.  These do Gaussian blurring using causal and anti-causal IIR filters in series. The result is an approximation to true Gaussian blurring that gets more accurate with increasing sigma, and which has runtime independent of sigma.
The approximation is as good as a 3-sigma truncated Gaussian when sigma>=1, with the
hitch that it can produce values slightly outside of the input range.  If
this proves to be a problem for anyone, clamping can be done inside the
implementation if the correct range is known, at little to no performance
cost.

I've supplied an SSE-optimized version and a normal floating point version,
both for images of floats only.  Using timings from my computer, I've attempted
to make convolveGaussian switch to the fastest method for a given sigma
(standard truncated-kernel blurring is faster for small sigma).

On my current hardware, blurring a 640x480 grayscale image takes 1.54 ms,
independent of sigma.  With the non-SSE IIR version, it takes 6.2 ms.

I've also fixed a couple of bugs in the existing truncated-kernel implementation.

Both algorithms now assume an infinite extension of the border rows/columns
instead of using reflection.