Support for new FFMPEG (tested against the latest SVN version as of the date of this commit).
img_convert has been depracated in favour of the GPL only libswscale.

FFMPEG will only be compiled in if you use --enable-gpl. If you distribute
programs which use libCVD with ffmpeg, then they have to be GPL'd.

libCVD has an array of colourspace conversion functions, so it may be possible
to use them instead of libswscale. That would enable the use of FFMPEG under
libCVD with the LGPL license.