Added a new video source for http video streams from some ethernet cameras. This class does not do initiate the http transfer, it only opens a file.  Use wget and a named pipe to grab live video. It has been added to open_video_source
Also added a very generic member to videobuffer, so that the lifetime
of arbitrary classes can be managed by a video buffer. This allows
the new ServerPushJpegBuffer to manage the istream assosciated with the
video. See open_video_source for an example of this. It will also allow
decorator buffers to optionally manage the lifetime of the class they are
decorating.