Modified calibrate to generalize the camera model.
The camera model is altered at compile time. Currently, it isn't templated
becuse of the use of some globals, and initialization like line 672.

Calibrate assumes that there are 4 non radial distortion parameters: the
optic axis, the focal length and the aspect ratio (stored as 2 focal lengths).
All remaining parameters are assumed to be radial distortion.

Tested with Quintic, Cubic and Harris models.