Fixed frame_pending() for VideoBuffer.
VideoBuffers can have one of three types: NotLive, Live and Flushable.

For NotLive, the semantics are the same. frame_pending() may be unset when the
last frame arrives (eg from DiskBuffer), and subsequent get_frame()s throw.

For Live, the video is from a live source (or behaves that way): frames arrive
at a fixed rate. However, frame_pending() always returns true.

Flushable buffers are live, but frame_pending() is useful.

This allows you to test if a buffer is flushable before attempting to flush it.
Alternatively, it allows you to determine if a buffer may have an end, before
testing if it has ended. VideoBuffer now has a flush method.