Added Bgrx type:
This type has blue, green, red, dummy. It differes from Bgra as it does not
pretend to have an alpha channel for the purposes of conversion.

NB: it appears that the Abgr type is misnamed. It looks much more like
BGRA.

Also, the connversion code from grey to luma-alpha, to RGB colour to RGBA
colour appears to require quadratic code in the number of types!

Image loading works on Bgrx types.