Image has value semantics and a move constructor
Note that changes propagate to various things especially the
video buffers (i.e. LocalVideoFrame based ones). Note that those
have only got move constructors, so they eat an image, and not
the obvious copy constructors. This is to make sure I didn't
introduce any inadvertant copies and since the API is so rarely used
directly (it's internal), I think the explicitness is good.