Add support for libUVC
This gives USB video camera support via a userland library. It's
somewhat unnecessary on Linux since v4l2 supports UVC probably
better, however, it means that the video capture ought to work
on other platforms too.

This required some final finishing on Image so that you can construct
proper memory managed images of dummy types. This was not possible before.
The code to do that is "interesting".